src/services/logging/requirements.txt
Pyro4==4.76
python-dotenv==0.10.3
pytz==2019.1
serpent==1.28
